requirements.txt
requirements.txt - Python Dependencies for ITNB Technical Assessment

## Core Dependencies
groundx>=0.3.0          # GroundX SDK for RAG
python-dotenv>=1.0.0    # Environment variable management
requests>=2.32.0        # HTTP client for API calls
beautifulsoup4>=4.12.0  # Web scraping

## Optional (for advanced features)
# openai>=1.3.0         # If using OpenAI for completions
# pydantic>=2.0.0       # Data validation (included via groundx)

Installation:
pip install -r requirements.txt

Or individually:
pip install groundx python-dotenv requests beautifulsoup4
